+++ src/main/java/com/ebock/service/TagService.java
@@ -60,6 +60,8 @@ public class TagService {
     @Path("/{id}")
     @RolesAllowed("admin")
     public Response delete(@PathParam("id") int id) {
+        if(tagMapper.getCountById(id) == 0)
+            throw new NotFoundException("Tag not found");
         tagMapper.delete(id);
         return Response.noContent().build();
     }
