+++ src/main/java/com/ebock/service/WearService.java
@@ -60,6 +60,8 @@ public class WearService {
     @Path("/{id}")
     @RolesAllowed("admin")
     public Response delete(@PathParam("id") int id) {
+        if(wearMapper.getCountById(id) == 0)
+            throw new NotFoundException("Wear not found");
         wearMapper.delete(id);
         return Response.noContent().build();
     }
